+++ src/data/joint_distributions/bootstrap.py
@@ -6,8 +6,6 @@ from . import gaussian  # noqa: F401
 from . import hypercube  # noqa: F401
 from . import mapped_joint_distribution  # noqa: F401
 from . import noisy_distribution  # noqa: F401
-from . import representor_distribution  # noqa: F401
-from . import paired_representor_distribution  # noqa: F401
 from . import staircase  # noqa: F401
 
 # Corresponding config modules
@@ -15,6 +13,4 @@ from .configs import gaussian as gaussian_config  # noqa: F401
 from .configs import hypercube as hypercube_config  # noqa: F401
 from .configs import mapped_joint_distribution as mapped_joint_distribution_config  # noqa: F401
 from .configs import noisy_distribution as noisy_distribution_config  # noqa: F401
-from .configs import representor_distribution as representor_distribution_config  # noqa: F401
-from .configs import paired_representor_distribution as paired_representor_distribution_config  # noqa: F401
 from .configs import staircase as staircase_config  # noqa: F401